raise warning level on msvc build for test_nua, cast -1 to unsigned and correct data types on several vars
darcs-hash:20061121202803-16063-2887c808e99920b42fd378c701b844b677557a71.gz

+++ win32/unistd.h
@@ -28,7 +28,7 @@
 #include <fcntl.h>
 #include <stdio.h>
 
-#define write(fd, buf, len) _write((fd), (buf), (len))
+#define write(fd, buf, len) _write((fd), (buf), (unsigned int)(len))
 #define read(fd, buf, len)  _read((fd), (buf), (len))
 #define close(fd)           _close((fd))
 #define mktemp(template)    _mktemp((template))
